Kitchen Layout for List of Orders
I have added some layout changes, these additions include a pair of
scroll-views within a linear layout that will be used to present the
orders that are input into the database. The list will pop up square
rows of a different color when the scroll-view is populated. I also
added the ingredients add in button since this is the only other
functionality needed for kitchen staff.

The next step will be adding java code to make all of this UI do the
things we need.