Added Survey Screen + More Detail to Coupons
The survey screen is now added and works in its entirety, although there
is currently no activity flow to reach the survey screen. For now, the
only way to access is by launching directly within Android Studio. The
user's selections are added to a new ParseObject. If they entered text
into the comments, that text will be pulled as well. Otherwise, the
string "None" will be sent to Parse for their comments. Aside from that,
I added the user to coupons as well as an "Unused" string into the
Status field.