What I have so far of making menu items toggle.
In "MenuAvailability.java" I have a function called "Toggle". This
function needs to take the current thing in the listview and change its
availability in parse to the opposite of what it is. Im having a hard
time doing it since I did it differently in the changing order status.

Anyway the xml's for that file are "activity_menu_availability.xml" and
"activity_availability_adaptor.xml"

Il try and figure it out later today but ive been up all night so ima
get a little sleep. Let me know if anyone has any ideas.